Refactor AssetRenditions to remove dependency cycle
