+++ core/src/main/java/com/adobe/aem/commons/assetshare/content/impl/datasources/AssetRenditionsDataSource.java
@@ -20,7 +20,7 @@
 package com.adobe.aem.commons.assetshare.content.impl.datasources;
 
 import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionDispatcher;
-import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditions;
+import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionDispatchers;
 import com.adobe.aem.commons.assetshare.util.DataSourceBuilder;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
@@ -57,12 +57,12 @@ public class AssetRenditionsDataSource extends SlingSafeMethodsServlet {
     private static final String PN_ADD_ASSET_RENDITION_DISPATCHER_TO_LABEL = "addAssetRenditionDispatcherToLabel";
 
     @Reference
-    private DataSourceBuilder dataSourceBuilder;
+    private transient DataSourceBuilder dataSourceBuilder;
 
     @Reference
-    private AssetRenditions assetRenditions;
+    private transient AssetRenditionDispatchers assetRenditionDispatchers;
 
-    private Cfg cfg;
+    private transient Cfg cfg;
 
     @Override
     protected final void doGet(SlingHttpServletRequest request, SlingHttpServletResponse response) {
@@ -82,7 +82,7 @@ public class AssetRenditionsDataSource extends SlingSafeMethodsServlet {
         final boolean addAssetRenditionDispatcherToLabel =
                 properties.get(PN_ADD_ASSET_RENDITION_DISPATCHER_TO_LABEL, cfg.add_assetrenditiondispatcher_to_label());
 
-        for (final AssetRenditionDispatcher assetRenditionDispatcher : assetRenditions.getAssetRenditionDispatchers()) {
+        for (final AssetRenditionDispatcher assetRenditionDispatcher : assetRenditionDispatchers.getAssetRenditionDispatchers()) {
 
             if (acceptsAssetRenditionDispatcher(allowedAssetRenditionTypes,
                     excludeAssetRenditionDispatchers,
@@ -144,7 +144,7 @@ public class AssetRenditionsDataSource extends SlingSafeMethodsServlet {
     }
 
     @Activate
-    protected void activate(AssetRenditionsDataSource.Cfg cfg) {
+    protected void activate(Cfg cfg) {
         this.cfg = cfg;
     }
