+++ core/src/main/java/com/adobe/aem/commons/assetshare/content/renditions/AssetRenditionDispatchers.java
@@ -1,0 +1,42 @@
+/*
+ * Asset Share Commons
+ *
+ * Copyright (C) 2019 Adobe
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.adobe.aem.commons.assetshare.content.renditions;
+
+import org.osgi.annotation.versioning.ProviderType;
+
+import java.util.List;
+
+/**
+ *
+ */
+@ProviderType
+public interface AssetRenditionDispatchers {
+    /**
+     * @return a list of all registered AssetRenditionDispatchers in the system ordered in Descending order by service.ranking.
+     */
+    List<AssetRenditionDispatcher> getAssetRenditionDispatchers();
+
+    /**
+     * Checks if the provided asset rendition name is a valid, registered asset rendition.
+     * @param name the asset rendition name.
+     * @return true if the name is registered as a valid asset rendition
+     */
+    default boolean isValidAssetRenditionName(String name) { return false; }
+}
