+++ core/src/main/java/com/adobe/aem/commons/assetshare/content/renditions/impl/AssetRenditionServlet.java
@@ -20,8 +20,8 @@
 package com.adobe.aem.commons.assetshare.content.renditions.impl;
 
 import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionDispatcher;
+import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionDispatchers;
 import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionParameters;
-import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditions;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
@@ -39,7 +39,10 @@ import javax.servlet.Servlet;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * /content/dam/foo.png.renditions/thumbnail/download/asset.rendition
@@ -68,16 +71,16 @@ public class AssetRenditionServlet extends SlingSafeMethodsServlet {
     public static final String SERVLET_EXTENSION = "renditions";
 
     @Reference
-    private AssetRenditions assetRenditions;
+    private transient AssetRenditionDispatchers assetRenditionDispatchers;
 
-    private Set allowedParameters = new HashSet();
+    private transient Set allowedParameters = new HashSet();
 
     public final void doGet(SlingHttpServletRequest request, SlingHttpServletResponse response) throws IOException, ServletException {
         try {
             final AssetRenditionParameters parameters = new AssetRenditionParameters(request);
 
             if (acceptsAssetRenditionParameters(parameters)) {
-                for (final AssetRenditionDispatcher assetRenditionDispatcher : assetRenditions.getAssetRenditionDispatchers()) {
+                for (final AssetRenditionDispatcher assetRenditionDispatcher : assetRenditionDispatchers.getAssetRenditionDispatchers()) {
                     if (acceptedByAssetRenditionDispatcher(assetRenditionDispatcher, parameters)) {
 
                         setResponseHeaders(response, parameters);
@@ -94,7 +97,7 @@ public class AssetRenditionServlet extends SlingSafeMethodsServlet {
                 response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Unsupported suffix parameters detected.");
             }
         } catch (IllegalArgumentException e) {
-            log.debug("Invalid request URL format for AssetRenditionServlet.", e);
+            log.warn("Invalid request URL format for AssetRenditionServlet.", e);
 
             response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR , "Invalid request URL format for AssetRenditionServlet.");
         }
@@ -102,7 +105,7 @@ public class AssetRenditionServlet extends SlingSafeMethodsServlet {
 
     protected boolean acceptedByAssetRenditionDispatcher(final AssetRenditionDispatcher assetRenditionDispatcher, final AssetRenditionParameters parameters) {
         if (assetRenditionDispatcher.getRenditionNames() == null ||
-                assetRenditions == null ||
+                assetRenditionDispatchers == null ||
                 StringUtils.isBlank(parameters.getRenditionName())) {
             return false;
         } else {
