+++ core/src/test/java/com/adobe/aem/commons/assetshare/content/renditions/impl/AssetRenditionServletTest.java
@@ -20,6 +20,7 @@
 package com.adobe.aem.commons.assetshare.content.renditions.impl;
 
 import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionDispatcher;
+import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionDispatchers;
 import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditionParameters;
 import com.adobe.aem.commons.assetshare.content.renditions.AssetRenditions;
 import com.adobe.aem.commons.assetshare.content.renditions.impl.dispatchers.StaticRenditionDispatcherImpl;
@@ -47,6 +48,7 @@ public class AssetRenditionServletTest {
         ctx.currentResource("/content/dam/test.png");
 
         ctx.registerService(AssetRenditions.class, new AssetRenditionsImpl());
+        ctx.registerService(AssetRenditionDispatchers.class, new AssetRenditionDispatchersImpl());
     }
 
     @Test
